refactor: moved functions to utils for more readability
